move files to better names
